Integration des dialogues : dialog configuration
