+++ app/src/main/res/menu/menu_configuration.xml
@@ -1,0 +1,13 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto">
+
+    <item
+        android:id="@+id/setting"
+        android:icon="@drawable/ic_iconsetting"
+        android:title="@string/menu_option"
+        app:showAsAction="always">
+
+    </item>
+
+</menu>
